Bind /usr/bin/env when cross building snapshots
The bindings allows to cross build snapshots when the target root filesystem
does not provide the /usr/bin/env command. This is the case at least for the
root filesystems provided by the ACI for ARM (v5, v6, v7, v7_uclibc). We can
now build v8 with snapshots also for these versions of ARM.