sh4: return a qNaN for the call to OS::nan_value()
The code is using the NAN macro defined in maths.h. While the C standard
require this value to be a qNaN, the current headers provided by the st40 tool
chain returns an sNaN.
In fact the header is the same on x86 and on sh4 while the signification of the
bit for sNaN vs. qNaN is inverted.