sh4: fix string-slices-regexp.js and the use of slice strings
The bug is due to one register to be scratched by the call to the runtime. In
fact, the caller saved register are saved before the call, but one of the
registers is changed between the push of these register and the real call. Ence
the old value is restored instead of the new one.