sh4: fix roots register number
Fixed reserved register for root pointer to be the same
as the roots register defined in assembler-sh4.h.
It was incorrectly set to r10 (as does ARM) while it should be
r12 on SH4.

This fixes a number of misaligned accesses in tests for the
LoadRoot() and CompareRoot() code generation.