sh4: rework MemOperand implementation and interface
Changes MemOperand interface to avoid ambiguities with
ARM implementation that use rn() as base and rm() as offset.
This is error prone as the SH4 implementation uses rm() as base.
Change the implementation and interface for SH4 such that
rn() is the base (same as ARM) and roffset() is the register offset.