sh4: speedup sh4 simulator by setting jump table static
Moves the jump table for the instruction decode from local
to a global static.
By having the jump table local, each time InstructionDecode()
was executed, the compiler copied the whole jump table onto
the stack frame, making the simulator spend 95% of its runtime
in memcpy().
Moving it globally speeds up dramatically the simulator with a
factor of more than x10.