sh4: implement code aging and support functions
Implements support for code aging which patches function
entry prologues.
The implementation is based on the ARM one but uses tricks from MIPS
in order to compute from the aging builtins the strt of the sequence
to patch.
On SH4 there is an additional issue with management of misaligned
sequence which makes the computation of the stub aging address more
complicated than on ARM or MIPS.
Refer to comments in codegen-sh4.cc.
Fixes cctest/test-heap-profiler/HeapSnapshotInternalReference.