sh4: activate compilation with debuggersupport=on
Activate debuggersupport=on for SH4 compilation.
This does not implement debug-sh4.cc, but just enables the
ENABLE_DEBUGGER_SUPPORT define.
This is necessary to allow correct backtrace when handling exception
as the FunctionGetDeferredName bultin is only active with debgger support.
As the debug-sh4.cc debugger support is actually not implemented,
this deactivates the debugger tests in cctest/test-debug.cc for SH4.
Fixes cctest/test-regressions-sh4/IssueEvalStackTrace.