sh4: augment first code page size for SH4
Augments first page size on SH4 to 540Kb instead of 416Kb on ARM.
This is needed such that code stubs fit in the first code page and
can be serialized.
The SH4 code generated is currently larger than the ARM code,
thus we have to adjust this.
Fixes cctest/test-spaces/SizeOfFirstPageIsLargeEnough.