sh4: implement canonicalization of NaNs
Implements the VFPCanonicalizeNaN function which was left as a no-op.
All NaNs stored to JS double arrays are canonicalized such that they
do not have the same value a the HoleNaN.
Note that for SH4 native execution the HoleNaN and canonical NaN should
be implemented as SH4 quiet NaN which is not the case currently.
Fixes mjsunit/regress/regress-2596.js.