sh4: implement debugger breakpoints support
Fixed and complete implementation of breakpoints support for
debuggers.
Re-activate tests that require debugger support.

The generated breakpoint slots have some constraints which
complicate the implementation for the sh4 architecture which have
alignment constraints. In particular the breakpoint slots must
be of fixed size and must not modify the initial alignment.